Fix: disklessset - Prevent rm if unmount failed

+++ bluebanquise-disklessset/diskless/livenet_module.py
@@ -453,10 +453,14 @@ class LivenetImage(Image):
         os.system('umount ' + self.MOUNT_DIRECTORY + '{proc,sys}')
 
         logging.debug('Executing \'umount ' + self.MOUNT_DIRECTORY + '\'')
-        os.system('umount ' + self.MOUNT_DIRECTORY)
+        umount_error_code = os.system('umount ' + self.MOUNT_DIRECTORY)
 
-        logging.debug('Executing \'rm -rf ' + self.MOUNT_DIRECTORY + '\'')
-        shutil.rmtree(self.MOUNT_DIRECTORY)
+        if umount_error_code == 0:
+            logging.debug('Executing \'rm -rf ' + self.MOUNT_DIRECTORY + '\'')
+            shutil.rmtree(self.MOUNT_DIRECTORY)
+        else:
+            logging.error('Could not umount ' + self.MOUNT_DIRECTORY + '. Please ensure no process is using this path.')
+            quit()
 
         # Create a squashfs backup (prevent failure)
         logging.debug('Executing \'mv ' + self.IMAGE_DIRECTORY + 'squashfs.img ' + self.IMAGE_DIRECTORY + 'squashfs.img.bkp\'')
@@ -751,7 +755,7 @@ def cli_menu():
                 cli_create_livenet_image_questions()
             elif action == 'Mount an existing livenet image':
                 cli_mount_livenet_image()
-            elif action == 'Unount an existing livenet image':
+            elif action == 'Unmount an existing livenet image':
                 cli_unmount_livenet_image()
             else:
                 cli_resize_livenet_image()
